requirements/base.txt
Django==1.7.1
bpython==0.13
django-braces==1.4.0
django-model-utils==2.0.3
djangorestframework==2.4.4
logutils==0.3.3
mutagen==1.26
watchdog==0.8.2
beets==1.3.8

